Refactor: rename webapi and restructure test helpers
