+++ source/Energinet.DataHub.MeteringPoints.EntryPoints.WebAPI/GridAreas/Create/Validation/Errors/PriceAreaCodeRuleError.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.EntryPoints.WebAPI.GridAreas.Create.Validation.Errors
+namespace Energinet.DataHub.MeteringPoints.EntryPoints.WebApi.GridAreas.Create.Validation.Errors
 {
     public class PriceAreaCodeRuleError : ValidationError
     {
