+++ source/Energinet.DataHub.MeteringPoints.IntegrationTests/Energinet.DataHub.MeteringPoints.IntegrationTests.csproj
@@ -47,7 +47,7 @@ limitations under the License.
     <ProjectReference Include="..\Energinet.DataHub.MeteringPoints.Client\Energinet.DataHub.MeteringPoints.Client.csproj" />
     <ProjectReference Include="..\Energinet.DataHub.MeteringPoints.EntryPoints.Common\Energinet.DataHub.MeteringPoints.EntryPoints.Common.csproj" />
     <ProjectReference Include="..\Energinet.DataHub.MeteringPoints.EntryPoints.Outbox\Energinet.DataHub.MeteringPoints.EntryPoints.Outbox.csproj" />
-    <ProjectReference Include="..\Energinet.DataHub.MeteringPoints.EntryPoints.WebAPI\Energinet.DataHub.MeteringPoints.EntryPoints.WebAPI.csproj" />
+    <ProjectReference Include="..\Energinet.DataHub.MeteringPoints.EntryPoints.WebApi\Energinet.DataHub.MeteringPoints.EntryPoints.WebApi.csproj" />
     <ProjectReference Include="..\Energinet.DataHub.MeteringPoints.Infrastructure\Energinet.DataHub.MeteringPoints.Infrastructure.csproj" />
     <ProjectReference Include="..\Energinet.DataHub.MeteringPoints.Messaging\Energinet.DataHub.MeteringPoints.Messaging.csproj" />
   </ItemGroup>
